Move DSpace-v6 code from refactor to main "dspace" package
